+++ CHANGELOG.md
@@ -19,6 +19,24 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ---
 
+## [0.6.0] - 2025-05-21
+
+### Added
+
+### Changed
+
+- Updated dev dependencies
+- Updated the generator code for complex data types (work-in-progress)
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
+
 ## [0.5.0] - 2025-05-17
 
 ### Added